requirements.txt
astroid==2.2.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
defusedxml==0.5.0
Django==2.1.7
django-admin==1.3.2
django-bootstrap==0.2.4
django-bootstrap4==0.0.7
django-dotenv==1.4.2
django-excel-response2==2.0.8
django-seed==0.1.9
django-six==1.0.4
djangorestframework==3.9.2
Faker==1.0.2
idna==2.8
isort==4.3.9
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
screen==1.0.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
text-unidecode==1.2
typed-ast==1.2.0
urllib3==1.24.1
wrapt==1.11.1
xlwt==1.3.0
